requirements.txt
openai>=1.35.0
pandas>=2.1.4
matplotlib>=3.8.0
seaborn>=0.13.0
tenacity>=8.2.3
requests>=2.31.0
python-dotenv>=1.0.0  